refactor: reemplazar diálogos de eliminación de tipos de personal por diálogos de desactivación
